[PROJ-56] feat: consulting tab 연결
