+++ src/App.tsx
@@ -2,7 +2,6 @@ import { NavermapsProvider } from 'react-naver-maps';
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import Layout from './components/template/Layout.tsx';
-import MyChatLayout from './components/template/MyChatLayout.tsx';
 import LoanDetail from './pages/LoanDetail.tsx';
 import LoanRecommend from './pages/LoanRecommned.tsx';
 import LoanReservation from './pages/LoanReservation/LoanReservation.tsx';
@@ -44,7 +43,6 @@ function App() {
             />
             <Route path="/live-chat" element={<ChatApp accessor="guest" />} />
             {/* 임시 */}
-            <Route path="/mychat" element={<MyChatLayout />} />
           </Routes>
         </Layout>
       </NavermapsProvider>
